Use tx-categorize for contract interaction labels
