+++ package.json
@@ -13,7 +13,6 @@
     "webpack:lavamoat:build": "yarn webpack:lavamoat -- --env production --no-cache",
     "webpack:lavamoat:generate": "yarn webpack:lavamoat --writeAutoPolicy -- --env production --no-cache",
     "foundryup": "yarn mm-foundryup",
-    "postinstall": "yarn webpack:clearcache && yarn foundryup",
     "env:e2e": "SEGMENT_HOST='https://api.segment.io' SEGMENT_WRITE_KEY='FAKE' yarn",
     "start": "yarn build:dev dev --apply-lavamoat=false --snow=false",
     "start:with-state": "node ./app/scripts/start-with-wallet-state.mjs",
@@ -394,6 +393,7 @@
     "@metamask/transaction-controller": "^62.8.0",
     "@metamask/transaction-pay-controller": "^10.2.0",
     "@metamask/tron-wallet-snap": "^1.19.1",
+    "@metamask/tx-categorize": "^0.1.0",
     "@metamask/user-operation-controller": "^40.0.0",
     "@metamask/utils": "^11.4.2",
     "@ngraveio/bc-ur": "^1.1.13",
