+++ ui/hooks/useTransactionDisplayData.js
@@ -10,6 +10,7 @@ import {
   selectERC20TokensByChain,
 } from '../selectors';
 import {
+  getContractInteractionLabel,
   getStatusKey,
   getTransactionTypeTitle,
 } from '../helpers/utils/transactions.util';
@@ -35,7 +36,6 @@ import { getIntlLocale } from '../ducks/locale/locale';
 import { NETWORK_TO_SHORT_NETWORK_NAME_MAP } from '../../shared/constants/bridge';
 import { calcTokenAmount } from '../../shared/lib/transactions-controller-utils';
 import { selectBridgeHistoryItemForTxMetaId } from '../ducks/bridge-status/selectors';
-import { getContractInteractionLabel } from '../helpers/utils/contract-interaction-heuristics';
 import { useI18nContext } from './useI18nContext';
 import { useTokenFiatAmount } from './useTokenFiatAmount';
 import { useUserPreferencedCurrency } from './useUserPreferencedCurrency';
